requirements.txt
redis~=5.1.1
faker==14.2.0
pandas~=2.2.2
pydantic==2.4.0
email-validator~=1.2.1
Flask==3.0.2
flask-sock==0.5.2
jproperties==2.1.1
mysql-connector-python==8.2.0
psycopg2-binary